bugfix/FT-01: replace init appService
