+++ src/js/templates/header/initHeader.js
@@ -1,26 +1,8 @@
 import createInitialHeaderMarkup from './createInitialHeaderMarkup';
-import { onLibraryClick } from './relinkingApi';
-import { appService } from '../../..';
+import { onLibraryClick, onSearchMovie } from './relinkingApi';
 
 // после инициализации блок header остается и рендеририться будет только разметка внутри
 // оставляю фиксированный блок для того, чтобы при переходе на другие страницы, не прыгала разметка
-
-const onSearchMovie = e => {
-  e.preventDefault();
-
-  appService.query = e.currentTarget.elements.searchQuery.value.trim();
-
-  const isEmptySearch = appService.searchQuery === '';
-
-  if (isEmptySearch) {
-    console.log('Пустий пошук');
-    //Зробити повідомлення про пустий пошук
-    return;
-  }
-
-  appService.resetPage();
-  appService.fetchSearchMovie().then(console.log);
-};
 
 const initHeader = () => {
   const headerMarkup = createInitialHeaderMarkup();
@@ -33,4 +15,4 @@ const initHeader = () => {
   searchInput.addEventListener('submit', onSearchMovie);
 };
 
-export { initHeader, onSearchMovie };
+export { initHeader };
